Moved all wgangp contents to a separate module.
