+++ wgangp/datamanager.py
